username to emailOrPhone sql function Integer type to BIGINT user Default image add

+++ src/main/resources/db/migration/C2__GET_USER_ADDRESS_BY_REGION_ID.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION get_region_address(regionId INTEGER)
+CREATE OR REPLACE FUNCTION get_region_address(regionId BIGINT)
     RETURNS TEXT
     LANGUAGE SQL AS
 $$
